+++ SensateIoT.SmartEnergy.Dsmr.Api/Properties/AssemblyInfo.cs
@@ -38,5 +38,5 @@ using SensateIoT.SmartEnergy.Dsmr.Api;
 //
 // You can specify all the values or you can default the Revision and Build Numbers
 // by using the '*' as shown below:
-[assembly: AssemblyVersion("1.6.0.0")]
-[assembly: AssemblyFileVersion("1.6.0.0")]
+[assembly: AssemblyVersion("1.7.0.0")]
+[assembly: AssemblyFileVersion("1.7.0.0")]