+++ SensateIoT.SmartEnergy.Dsmr.Data/DTO/DeviceCapability.cs
@@ -1,0 +1,8 @@
+﻿namespace SensateIoT.SmartEnergy.Dsmr.Data.DTO
+{
+	public class DeviceCapability
+	{
+		public bool HasSolarCells { get; set; }
+		public bool HasGasMeter { get; set; }
+	}
+}