+++ SensateIoT.SmartEnergy.Dsmr.Data/Models/DeviceCapability.cs
@@ -1,0 +1,8 @@
+﻿namespace SensateIoT.SmartEnergy.Dsmr.Data.Models
+{
+	public class DeviceCapability
+	{
+		public bool HasSolarCells { get; set; }
+		public bool? HasGasMeter { get; set; }
+	}
+}