+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 22-12-2021
+### Added
+- An endpoint for average energy usage aggregates
+- Swagger documentation
+
+### Updated
+- Authorization errors
+- Project dependencies
+
 ## [1.6.0] - 18-06-2021
 ### Added
 - Sensor availability per device